serial: Fix warning about comparing a signed and unsigned int

+++ serial.c
@@ -275,7 +275,7 @@ static ssize_t serial_write_data(struct iio_context_pdata *pdata,
 	if (ret < 0) {
 		IIO_ERROR("sp_blocking_write returned %i\n", (int) ret);
 		return ret;
-	} else if (ret < len) {
+	} else if ((size_t) ret < len) {
 		IIO_ERROR("sp_blocking_write has timedout\n");
 		return -ETIMEDOUT;
 	}
